services/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "echo-mvp-api"
version = "0.1.0"
description = "Echo MVP FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.35",
  "alembic>=1.13.2",
  "python-multipart>=0.0.9",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "bcrypt<4"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.8.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
